requirements.txt
langchain_community
langchain_core
langchain_groq
fastapi
pypdf
dotenv
python-docx
reportlab
pydantic
streamlit